Refactored Extended Demographics into separate classes/logic
